requirements.txt
kafka-python==2.0.2
tensorflow==2.11.0
Pillow
numpy==1.23.4